Requeue the request if API mapping isn't found (#96)
* Requeue the request if API mapping isn't found

This should ensure that the template-sync keeps retrying on Policies
where the specific policy type CRD (eg ConfigurationPolicy) hasn't been
created on the cluster yet. Previously, the template sync would try a
few times, but then these policies would not get retried until the cache
was fully refreshed, or the controller restarted.

Note on requeue behavior: it is not instantly reconciled again. From an
experiment, it looks like it tries again after 5 seconds about a dozen
times, then backs off exponentially to a 16 minute, 40 second wait. That
maximum wait could be configured through the RateLimiter controller
option (currently, this uses the default).

Refs:
 - https://github.com/stolostron/backlog/issues/21640

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update CRD to apiextensions.k8s.io/v1

The latest KinD version removed the deprecated v1beta1 definition for
CRDs.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Fix test for CRD not found events

Previously, only two events would ever be created in this case, but with
the bugfix it will create more events while it checks if the CRD was
installed. So, the test needs to delete the policy first, then delete
*all* the related events.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>