Emit noncompliance events on template errors
In a variety of error situations, the template sync should mark the
policy as non-compliant by emitting an event of the proper format. Then,
after an error is resolved, it should signal the relevant policy
controller to re-send a [non]compliance event for the template. It does
this by clearing the current complianceState in the status of the
template object.

While improving error flows in general, other potential bugs have been
fixed: for example, previously if there was an issue with one of the
policy templates, none of the later ones in the list would be created
or updated. Now, all templates are handled.

Also adds a bunch of tests for the new behavior, the bug described
above, and code coverage in general.

Refs:
 - https://github.com/stolostron/backlog/issues/24750

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit b07a80b64cfe70e36741630bdaa3e2e6cfb9443e)