Handle when the template sync encounters a conflict
When there is a conflict, the reconcile should be retried rather than
emit a policy violation.

This a cherry-pick from:
https://github.com/open-cluster-management-io/governance-policy-framework-addon/commit/5711c68990c3d57f65422dfb6093c7c0694c8035

This change is necessary in case the consolidated sync controller
project is not complete by the next release.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit d11aa59efdada8316c08465c8c5566915fe6da51)